feat: TaskQueue
Finish tasks asyncronously, but sequencially mark done/timeout/error.

Queue tasks of any key/value types and run them as soon as they are added, then trigger `TaskDone` when the top task is done. once the top is done, it will check the next until it reaches a `Pending` one so it stops and waits for it to finish